Add support the SMB100B driver
